requirements.txt
# Core AI and Agent Framework
langchain-core==0.3.12
langchain-community==0.3.7
langgraph==0.2.34
langchain-ollama==0.2.0

# Machine Learning and Embeddings
sentence-transformers==3.1.1
torch>=2.0.0
numpy>=1.24.0

# Data and Dataset Handling
datasets==3.0.1
huggingface-hub>=0.19.0

# Web Interface
streamlit>=1.28.0

# Search and Tools
duckduckgo-search>=6.2.0

# Utilities
typing-extensions>=4.8.0 